+++ plugins/modules/elastic_endpoint_security.py
@@ -10,42 +10,6 @@ from ansible.module_utils.basic import _ANSIBLE_ARGS, AnsibleModule
 import json
 
 try:
-  from ansible_collections.expedient.elastic.plugins.module_utils.elastic_integration import Integration
-except:
-  import sys
-  import os
-  util_path = new_path = f'{os.getcwd()}/plugins/module_utils'
-  sys.path.append(util_path)
-  from elastic_integration import Integration
-
-try:
-  from ansible_collections.expedient.elastic.plugins.module_utils.elastic_agentpolicy import AgentPolicy
-except:
-  import sys
-  import os
-  util_path = new_path = f'{os.getcwd()}/plugins/module_utils'
-  sys.path.append(util_path)
-  from elastic_agentpolicy import AgentPolicy
-  
-try:
-  from ansible_collections.expedient.elastic.plugins.module_utils.elastic_pkgpolicy import PkgPolicy
-except:
-  import sys
-  import os
-  util_path = new_path = f'{os.getcwd()}/plugins/module_utils'
-  sys.path.append(util_path)
-  from elastic_pkgpolicy import PkgPolicy
-
-try:
-  from ansible_collections.expedient.elastic.plugins.module_utils.elastic_rules import Rules
-except:
-  import sys
-  import os
-  util_path = new_path = f'{os.getcwd()}/plugins/module_utils'
-  sys.path.append(util_path)
-  from elastic_rules import Rules
-  
-try:
   from ansible_collections.expedient.elastic.plugins.module_utils.ece import ECE
 except:
   import sys
@@ -89,7 +53,7 @@ class SecurityBaseline(Kibana):
         
     def create_securityctrl_baseline_settings(self,pkg_policy_object):
         ################ Checking and creating package policy associated with Integration
-        pkg_policy_action = PkgPolicy(self.module)
+        pkg_policy_action = Kibana(self.module)
         try:
           pkg_policy_object['package']
         except:
@@ -117,7 +81,7 @@ class SecurityBaseline(Kibana):
           pkg_policy_info = pkg_policy_update
         
         elif pkg_policy_object['package']['title'] == 'Prebuilt Security Detection Rules' and self.prebuilt_rules_activate == True and self.module.check_mode == False:
-              rules_action = Rules(self.module)
+              rules_action = Kibana(self.module)
               #SecurityRules = rules_action.activating_all_rules(self,50)
               pkg_policy_info = rules_action.activate_rule(50,'Endpoint Security')
         
@@ -129,33 +93,36 @@ class SecurityBaseline(Kibana):
 def main():
 
     module_args=dict(   
-        host=dict(type='str',default='id'),
+        host=dict(type='str',Required=True),
         port=dict(type='int', default=9243),
-        username=dict(type='str', default='test1'),
-        password=dict(type='str', no_log=True, default='test1'),   
+        username=dict(type='str', Required=True),
+        password=dict(type='str', no_log=True, Required=True),   
         verify_ssl_cert=dict(type='bool', default=True),
         agent_policy_id=dict(type='str'),
         agent_policy_name=dict(type='str'),
-        integration_name=dict(type='str', default='Int Name'),
-        pkg_policy_name=dict(type='str', default='Int Pkg Name'),
-        pkg_policy_desc=dict(type='str', default='Int Pkg Desc'),
-        endpoint_security_antivirus=dict(type='bool', default=False),
-        prebuilt_rules_activate=dict(type='bool', default=False)
+        integration_name=dict(type='str', Required=True),
+        pkg_policy_name=dict(type='str', Required=True),
+        pkg_policy_desc=dict(type='str'),
+        endpoint_security_antivirus=dict(type='bool', default=True),
+        prebuilt_rules_activate=dict(type='bool', default=True),
+        state=dict(type='str', default='present')
     )
     argument_dependencies = []
         #('state', 'present', ('enabled', 'alert_type', 'conditions', 'actions')),
         #('alert-type', 'metrics_threshold', ('conditions'))
     
     module = AnsibleModule(argument_spec=module_args, required_if=argument_dependencies, supports_check_mode=True)
+    state = module.params.get('state')
     
     if module.check_mode:
         results['changed'] = False
     else:
         results['changed'] = True
         
-    agent_policy_action = AgentPolicy(module)
+    agent_policy_action = Kibana(module)
+    
     if not module.params.get('agent_policy_id'):
-      agency_policy_object = agent_policy_action.get_agent_policy_id_byname()
+      agency_policy_object = agent_policy_action.get_agent_policy_byid(agent_policy_action.get_agent_policy_id_byname())
     else:
       agency_policy_object = agent_policy_action.get_agent_policy_byid()
     try:
@@ -166,7 +133,7 @@ def main():
       results['changed'] = False
       module.exit_json(**results)
     
-    ElasticIntegration = Integration(module)
+    ElasticIntegration = Kibana(module)
     if module.params.get('integration_name'):
       integration_object = ElasticIntegration.check_integration(module.params.get('integration_name'))
     else:
@@ -175,23 +142,27 @@ def main():
       module.exit_json(**results)
     
     if not integration_object:
-      results['integration_status'] = 'Integration name is not a valid'
+      results['integration_status'] = 'Integration name is not valid'
       results['changed'] = False
       module.exit_json(**results)
     
-    pkg_policy_action = "create"
-    pkg_policy = PkgPolicy(module)
-    if pkg_policy_action == "create":
+    pkg_policy = Kibana(module)
+    if state == "present":
       pkg_policy_object = pkg_policy.get_pkg_policy(agent_policy_id)
       if pkg_policy_object:
         results['pkg_policy_status'] = "Integration Package found, No package created"
         results['changed'] = False
         results['pkg_policy_object'] = pkg_policy_object
-      else:    
-        pkg_policy_object = pkg_policy.create_pkg_policy(agent_policy_id, integration_object)
-        results['pkg_policy_status'] = "No Integration Package found, Package Policy created"
-        results['pkg_policy_object'] = pkg_policy_object
-            
+      else:
+        if module.check_mode == False:    
+          pkg_policy_object = pkg_policy.create_pkg_policy(agent_policy_id, integration_object)
+          results['pkg_policy_status'] = "No Integration Package found, Package Policy created"
+          results['changed'] = True
+          results['pkg_policy_object'] = pkg_policy_object
+        else:
+          results['pkg_policy_status'] = "No Integration Package found, Package Policy not created becans check_mode is set to true"
+          results['changed'] = False
+              
     ElasticSecurityBaseline = SecurityBaseline(module)
     if integration_object['title'] == 'Endpoint Security' or integration_object['title'] == 'Prebuilt Security Detection Rules':
       updated_pkg_policy_object = ElasticSecurityBaseline.create_securityctrl_baseline_settings(pkg_policy_object)
